requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
webrtcvad-wheels==2.0.14
openai>=1.50.0
python-dotenv==1.0.0
asyncpg==0.29.0
